Clean up RE data validation
