+++ radix-engine/src/engine/mod.rs
@@ -1,13 +1,15 @@
 mod allocator;
-mod env;
-mod error;
-mod loader;
+mod data_validator;
 mod process;
 mod track;
+mod wasm_env;
+mod wasm_loader;
+mod wasm_validator;
 
 pub use allocator::IdAllocator;
-pub use env::{EnvModuleResolver, KERNEL_INDEX, KERNEL_NAME};
-pub use error::RuntimeError;
-pub use loader::{instantiate_module, parse_module, validate_module};
+pub use data_validator::validate_data;
 pub use process::{Invocation, Process};
 pub use track::Track;
+pub use wasm_env::{EnvModuleResolver, KERNEL_INDEX, KERNEL_NAME};
+pub use wasm_loader::instantiate_module;
+pub use wasm_validator::{parse_module, validate_module};
