+++ radix-engine/src/model/mod.rs
@@ -1,17 +1,29 @@
 mod actor;
 mod bucket;
 mod component;
+mod error;
 mod lazy_map;
 mod nft;
 mod package;
+mod receipt;
 mod resource_def;
+mod transaction;
+mod validated_data;
+mod validated_transaction;
 mod vault;
 
 pub use actor::Actor;
 pub use bucket::{Bucket, BucketError, BucketRef, LockedBucket, Supply};
 pub use component::{Component, ComponentError};
+pub use error::{
+    DataValidationError, RuntimeError, TransactionValidationError, WasmValidationError,
+};
 pub use lazy_map::{LazyMap, LazyMapError};
 pub use nft::{Nft, NftError};
 pub use package::Package;
+pub use receipt::Receipt;
 pub use resource_def::{ResourceDef, ResourceDefError};
+pub use transaction::{Instruction, Transaction};
+pub use validated_data::ValidatedData;
+pub use validated_transaction::{ValidatedInstruction, ValidatedTransaction};
 pub use vault::{Vault, VaultError};
