+++ radix-engine/src/utils/mod.rs
@@ -1,5 +1,3 @@
-mod decoder;
 mod formatter;
 
-pub use decoder::decode_data;
 pub use formatter::{format_data, format_data_with_ledger};
